python-backend/requirements.txt
# Core Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
python-dotenv>=1.0.0
pydantic>=2.10.0

# AI Engine - Multi-Provider Support
openai>=1.57.0
cohere>=5.11.0
google-generativeai>=0.8.0
requests>=2.31.0

# Voice Engine
SpeechRecognition>=3.10.0
pyttsx3>=2.90
# pyaudio - Install separately: pip install pyaudio (or pipwin install pyaudio on Windows)

# Vision Engine
mss>=9.0.0
pillow>=10.4.0
numpy>=1.26.0
opencv-python>=4.9.0
pytesseract>=0.3.10

# Automation Engine
pyautogui>=0.9.54
keyboard>=0.13.5
pyperclip>=1.8.2

# Utilities (already included above in AI Engine section)
